✨ Add usage of a specific endpoint port
When pulling cluster from Consul, the port exposed is not
necessarily the one that must be used in order to authenticate
on the cluster. This feature allows to set a different endpoint
pot to authenticate on the ElasticSearch cluster.

By default, the endpoint port is the one exposed in Consul.

Signed-off-by: f.bouzghaia <f.bouzghaia@criteo.com>